* commit '4ab79ba72139765353af1df014af0e325b2b9806': (316 commits)
  REPORT-28830 提示优化
  REPORT-28894 多次切换远程的问题 && xml不可读处理
  fix mac下不用awt的库读写
  REPORT-28873
  REPORT-28792 表单添加tab块问题
  fix dock name
  fix 遗漏
  REPORT-28846 && REPORT-28830 && REPORT-27109
  REPORT-27106 && REPORT-28229
  CHART-13052 为仪表盘坐标轴轴标签htmlPane分配parentPane
  CHART-13064 修改含有自动button的颜色选择器高度
  fix import
  REPORT-27106 && REPORT-27144
  CHART-13065 调整字符间距
  CHART-13063 修改试管仪表盘布局文案
  REPORT-28186 mac出现短暂原生java图标问题
  REPORT-28368 死锁问题 常量不需要每次读取
  REPORT-28257 && REPORT-28257
  合代码冲突
  REPORT-28368 死锁问题 常量不需要每次读取
  ...

+++ designer-realize/src/main/java/com/fr/grid/CellElementPainter.java
@@ -9,7 +9,7 @@ import com.fr.report.core.PaintUtils;
 import com.fr.report.elementcase.ElementCase;
 
 public class CellElementPainter {
-	public void paintBackground(Graphics2D g2d, ElementCase report, CellElement ce, int width, int height) {
+	public void paintBackground(Graphics2D g2d, ElementCase report, CellElement ce, double width, double height) {
 		Style.paintBackground(g2d, ce.getStyle(), width, height);
 	}
